fix(work): apply file-path linkification inside a2ui Markdown component too
The fence-stripped Hermes path emits a synthetic a2ui surface containing
a Markdown component holding the agent prose. That Markdown is rendered
by apps/web/src/a2ui/components.tsx, not by work-screen.tsx's
RunTimeline — so my earlier preprocessing only ran on the RunTimeline
branch and never touched the surface-wrapped prose. As a result the PDF
path stayed as a plain <code> instead of a clickable link.

Extract linkifyWorkspacePaths into apps/web/src/lib/ so both renderers
share it, and call it inside the a2ui Markdown registration. Same regex,
same /api/work/files target. work-screen.tsx now imports the shared
helper instead of duplicating.
