+++ apps/web/src/a2ui/components.tsx
@@ -10,6 +10,7 @@
 
 import ReactMarkdown from "react-markdown";
 import remarkGfm from "remark-gfm";
+import { linkifyWorkspacePaths } from "@/lib/linkify-workspace-paths";
 import { registerComponent, renderChildren } from "./renderer";
 import type { RenderContext } from "./renderer";
 import type { A2UIComponent } from "./types";
@@ -37,7 +38,9 @@ registerComponent("Markdown", (comp: A2UIComponent) => {
   const props = comp as unknown as MarkdownProps & { id: string };
   return (
     <div key={props.id} className="work-markdown">
-      <ReactMarkdown remarkPlugins={[remarkGfm]}>{props.text}</ReactMarkdown>
+      <ReactMarkdown remarkPlugins={[remarkGfm]}>
+        {linkifyWorkspacePaths(props.text)}
+      </ReactMarkdown>
     </div>
   );
 });
